requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
dj-database-url==1.2.0
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
# psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
whitenoise==6.3.0
